backend/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
pydantic-settings>=2.0.0  # For settings management
python-multipart
coqui-tts
gtts>=2.5.0
pydub>=0.25.1
google-cloud-texttospeech>=2.16.0

# Database & ORM
sqlalchemy>=2.0
alembic>=1.12
psycopg2-binary>=2.9.0  # PostgreSQL adapter for Python

# Authentication & Security
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
PyJWT>=2.8.0

# Password hashing - using bcrypt directly (no passlib dependency)
bcrypt>=4.0.0

# Optional translator package - install if you want translation support
# indic-trans2
